Use ints instead of longs for 32 bit integers to fix x86_64 support
I could have used C99 types but CERTAIN PEOPLE don't have C99

+++ src/common.c
@@ -58,7 +58,7 @@ uchar	KeyDescTable[] = {
 
 char *KeySelect(uint KeyIndex)
 {
-	ulong *pKeyDescTable = (ulong*)KeyDescTable;
+	uint *pKeyDescTable = (uint*)KeyDescTable;
 	char *pRet;
 
 	if (pKeyDescTable[1] <= 0x1000)
